子账号功能

+++ buzmgt/src/main/java/com/wangge/buzmgt/region/web/RegionController.java
@@ -289,7 +289,7 @@ public class RegionController {
 		} else {
 			if (region.getChildren().size() > 0) {
 				for (Region reg : region.getChildren()) {
-					if (maxid < Integer.parseInt(reg.getId())) {
+					if (maxid < Long.parseLong(reg.getId())) {
 						maxid = Long.parseLong(reg.getId());
 					}
 				}
